Refactors package name
